fix(BudgetSection): Query + preloading
